add wallet events

+++ src/main/kotlin/co/akoot/plugins/bluefox/api/events/PlayerDepositEvent.kt
@@ -3,4 +3,4 @@ package co.akoot.plugins.bluefox.api.events
 import co.akoot.plugins.bluefox.api.economy.Coin
 import org.bukkit.entity.Player
 
-class PlayerDepositEvent(val player: Player, val coin: Coin, val amount: Double): FoxEventCancellable()
+class PlayerDepositEvent(val player: Player, val coin: Coin, val amount: Int): FoxEventCancellable()
